Shop and Crate Menus done, Generation and Scarcity done, Merchants added

+++ Hourglass Sands/Assets/Scripts/ItemScripts/PlayerInventory.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,12 +8,12 @@ public class PlayerInventory : InventoryComponent
 {
     public float currency;
     public UnityEvent e_CurrencyUpdated;
-    public void PurchaseItem(int id, int amount, int costPer) {
+    public void PurchaseItem(int id, int amount, float costPer) {
         currency -= amount * costPer;
         e_CurrencyUpdated.Invoke();
         AddItem(id, amount);
     }
-    public void SellItem(int id, int amount, int costPer) {
+    public void SellItem(int id, int amount, float costPer) {
         currency += amount * costPer;
         e_CurrencyUpdated.Invoke();
         RemoveItem(id, amount);
